requirements.txt
agno==0.1.0
groq==0.4.0
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7
